Fixed emit of migrate result (load still wrong).
