+++ src/message/migfra/result.cpp
@@ -8,6 +8,9 @@
 
 #include <fast-lib/message/migfra/result.hpp>
 
+// Alias merge_node function
+const auto &merge_node = fast::yaml::merge_node;
+
 namespace fast {
 namespace msg {
 namespace migfra {
@@ -28,7 +31,10 @@ YAML::Node Result_container::emit() const
 {
 	YAML::Node node;
 	node["result"] = title;
-	node["list"] = results;
+	if (title == "vm migrated")
+		merge_node(node, results.at(0).emit());
+	else
+		node["list"] = results;
 	if (id != "")
 		node["id"] = id;
 	return node;
