+++ src/serializable.cpp
@@ -18,4 +18,19 @@ namespace fast
 	{
 		load(YAML::Load(str));
 	}
+
+	namespace yaml {
+	
+		void merge_node(YAML::Node &lhs, const YAML::Node &rhs)
+		{
+			for (const auto &node : rhs) {
+				std::string tag = YAML::Dump(node.first);
+				if (!lhs[tag]) {
+					lhs[tag] = rhs[tag];
+				}
+			}
+		}
+	
+	}
+
 }
